ms-reserve/requirements.txt
flask == 2.3.3
python-dotenv == 1.0.0
flask-SQLAlchemy == 3.0.5
Flask-Migrate == 4.0.4
psycopg[binary]
psycopg2 == 2.9.7
mysqlclient == 2.2.4
flask-marshmallow==0.15.0
requests==2.31.0